refactor : 입력에 따른 패키지 분리
