+++ src/main/java/roomescape/RoomescapeConsoleRunner.java
@@ -5,8 +5,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.web.servlet.WebMvcAutoConfiguration;
 import org.springframework.context.annotation.Profile;
-import roomescape.controller.ReservationConsoleController;
-import roomescape.controller.ReservationTimeConsoleController;
+import roomescape.controller.console.ReservationConsoleController;
+import roomescape.controller.console.ReservationTimeConsoleController;
 import roomescape.view.ConsoleInputView;
 import roomescape.view.ConsoleOutputView;
 import roomescape.view.ReservationCommand;
