+++ src/main/java/roomescape/controller/console/ReservationTimeConsoleController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.console;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Controller;
